+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # PyMagento Changelog
 
+## Unreleased
+
+* `save_product`: add `log_response=False` to disable the log of the Magento response
+* `get_order` and `save_attribute`: remove the `throw` parameter; the default was `True` but even if one passed `False`
+  the method would still throw when trying to JSON-decode the error responses.
+* Simplify some code using [`api-session`](https://github.com/bixoto/api-session) 1.3.4+
+
 ## 1.9.3 (2024/01/08)
 
 * Add `get_cart`